move some responsibility for cache to db-api
